Update TicTacToe.java
Fixed the bug where if the user starting second and won, the CLI displayed "You lost."
Fixed the redundancy in the playing loop by creating the following methods:

    winnerDeclare

    userInput

    whoPlayFirst

    playingMethod

    playerMove

    gameOrRoundWon

Handled the draw case with the drawDetector method.
Added a sleeping time for the bot move to improve the UX, with a default value of 1000 milliseconds. The time can be changed using the advanceSettings method.

Made the second mode more dynamic by allowing the user to choose how many rounds are needed to win the game.
Gave the user the ability to choose the bot name.